125_выносит анимацию инкремента из SelectMovieFragment в отдельный класс
