+++ app/src/main/java/com/davay/android/feature/selectmovie/presentation/animation/IncrementAnimation.kt
@@ -1,0 +1,7 @@
+package com.davay.android.feature.selectmovie.presentation.animation
+
+import android.view.View
+
+interface IncrementAnimation {
+    fun animate(view: View, onAnimationEnd: () -> Unit)
+}
